rename redis_server_starter
