+++ pyabc/sampler/redis_eps/__init__.py
@@ -2,6 +2,6 @@
 
 from .sampler import RedisEvalParallelSampler
 from .sampler_static import RedisStaticSampler
-from .redis_sampler_server_starter import (
+from .server_starter import (
     RedisEvalParallelSamplerServerStarter,
     RedisStaticSamplerServerStarter)
